Created test file to turn the water pump on when the soil moisture reading is "too dry".
